:recycle: (pc-15004) move CountryPicker to identityCheck components
