+++ src/features/auth/signup/PhoneValidation/SetPhoneNumberDeprecated.tsx
@@ -11,6 +11,7 @@ import { Paragraphe } from 'features/auth/components/signupComponents'
 import { useAppSettings } from 'features/auth/settings'
 import { SignupStep } from 'features/auth/signup/enums'
 import { useSendPhoneValidationMutation } from 'features/identityCheck/api/api'
+import { CountryPicker, METROPOLITAN_FRANCE } from 'features/identityCheck/components/countryPicker'
 import { useIdentityCheckContext } from 'features/identityCheck/context/IdentityCheckContextProvider'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
@@ -29,8 +30,6 @@ import { useModal } from 'ui/components/modals/useModal'
 import { Close } from 'ui/svg/icons/Close'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
 import { Form } from 'ui/web/form/Form'
-
-import { CountryPicker, METROPOLITAN_FRANCE } from './components'
 
 const TIMER = 60
